+++ NEWS.md
@@ -1,5 +1,6 @@
 # TruncExpFam (development version)
 
+* Introduced sampling directly from the truncated distribution (issue #78)
 * Added a `NEWS.md` file to track changes to the package.
 * Added support for simple numeric-class vectors on `mlEstimationTruncDist()` (issue #95)
 * Improved linting (issue #83)